Finished task 3
