+++ product-service/lambda-handlers/getAllProducts.js
@@ -1,0 +1,22 @@
+import products from '../products.json';
+import ProductsService from '../services/products.service';
+
+export const getAllProducts = async event => {
+  try{
+    const productsServiceInstance = new ProductsService( products );
+    const allProducts = await productsServiceInstance.getProductsList();
+    
+    return {
+      statusCode: 200,
+      body: JSON.stringify( allProducts )
+    }
+  }
+  catch(error) {
+    console.error(JSON.stringify(error));
+    
+    return {
+      statusCode: 500,
+      body: "Something went wrong!"
+    };
+  }
+};
