+++ product-service/lambda-handlers/index.js
@@ -1,0 +1,8 @@
+import { from } from "webpack-sources/lib/CompatSource";
+import { getProductById } from './getProductById';
+import { getAllProducts } from './getAllProducts';
+
+export {
+    getProductById,
+    getAllProducts
+};
